区块调整

+++ Assets/Scripts/Save And Load/MapData.cs
@@ -10,13 +10,13 @@ public class MapData {
 
     //瓦片地图数据索引
     [Key(0)]
-    public int[,][,,] chunkDatas;
+    public long[,][,,] chunkDatas;
 
     [Key(1)]
     public long[,,] tileDatas;//瓦片Id
     
 
-    public MapData(int[,][,,] chunkDatas, long[,,] tileDatas) {
+    public MapData(long[,][,,] chunkDatas, long[,,] tileDatas) {
         this.chunkDatas = chunkDatas;
         this.tileDatas = tileDatas;
     }
